# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meteoro Task API",
        "description": "API документация для интернет-магазина",
        "contact": {
            "email": "admin@meteoro.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Получить список всех заказов (админ)",
                "description": "Возвращает список всех заказов с возможностью фильтрации по статусу",
                "operationId": "adminGetOrders",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу заказа",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "processing",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 59999.98
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+7 (999) 123-45-67"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "ул. Ленина, д. 1"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Иван Иванов"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "ivan@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Получить заказ по ID (админ)",
                "description": "Возвращает детальную информацию о заказе",
                "operationId": "adminGetOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 59999.98
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+7 (999) 123-45-67"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "ул. Ленина, д. 1"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Иван Иванов"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}/status": {
            "patch": {
                "tags": [
                    "Admin/Orders"
                ],
                "summary": "Обновить статус заказа (админ)",
                "description": "Изменяет статус заказа",
                "operationId": "adminUpdateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "processing",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "processing"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "processing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}